requirements.txt
appdirs==1.4.4
atomicwrites==1.4.0
attrs==20.2.0
black==20.8b1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
codecov==2.1.10
coverage==5.3
coveralls==2.1.2
docopt==0.6.2
idna==2.10
iniconfig==1.1.1
more-itertools==8.5.0
mypy-extensions==0.4.3
numpy==1.19.2
packaging==20.4
pandas==1.1.3
pathspec==0.8.0
pluggy==0.13.1
py==1.9.0
pympi-ling @ git+https://github.com/sarpu/pympi.git@bfc9d58e22ff85613a5eb5cdd13dc0aa226786dd
pyparsing==2.4.7
pytest==6.1.1
pytest-cov==2.10.1
python-dateutil==2.8.1
pytz==2020.1
pyvyu==0.3.6
regex==2020.10.15
requests==2.24.0
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
